Add unit test for new isTurnFirstOfRound
In addition to adding the test, extracted some utilities for making
testing easier.

# Conflicts:
#	src/composables/useGameLogic.test.ts
#	src/test-factories/index.ts
