+++ src/test-factories/createMockGlobalI18n.ts
@@ -1,0 +1,9 @@
+import { vi } from 'vitest';
+
+/* ========================================================================== */
+
+export function createMockGlobalI18n(): void {
+	vi.mock('@/i18n', () => ({
+		useGlobalI18n: () => ({ t: (key: string) => key })
+	}));
+}
